+++ blacklist-as.sh
@@ -30,13 +30,13 @@ for i in 45102 200373 198571 4134 4837 136907 55990 63727 9808; do
     grep "^$i," rpki.cidr | grep -v '::' >> $rpkideny_file
 done
 
-/sbin/iptables -D NEEDRESTRICT -m set --match-set asn-blocklist-v4 src -j DROP >/dev/null 2>&1
+#/sbin/iptables -D NEEDRESTRICT -m set --match-set asn-blocklist-v4 src -j DROP >/dev/null 2>&1
+/sbin/iptables -D INPUT -m set --match-set asn-blocklist-v4 src -j DROP >/dev/null 2>&1
 /sbin/ipset destroy asn-blocklist-v4 >/dev/null 2>&1
 
 /sbin/ipset create asn-blocklist-v4 hash:net comment
 
 awk -F, '{print "add asn-blocklist-v4 "$2" comment "$1}' $rpkideny_file | /sbin/ipset restore
 
-/sbin/iptables -I NEEDRESTRICT -m set --match-set asn-blocklist-v4 src -j DROP
-#/sbin/iptables -I INPUT -m set --match-set asn-blocklist-v4 src -j DROP
-
+#/sbin/iptables -I NEEDRESTRICT -m set --match-set asn-blocklist-v4 src -j DROP
+/sbin/iptables -I INPUT -m set --match-set asn-blocklist-v4 src -j DROP