+++ blacklist-countries.sh
@@ -19,13 +19,13 @@ for i in CN KR RU; do
     awk -F"|" "/${i}/"'{print "add countries-blocklist-v4 "$2" comment "$1}' ripe.cidr >> $ripedeny_file
 done
 
-/sbin/iptables -D NEEDRESTRICT -m set --match-set countries-blocklist-v4 src -j DROP >/dev/null 2>&1
+#/sbin/iptables -D NEEDRESTRICT -m set --match-set countries-blocklist-v4 src -j DROP >/dev/null 2>&1
+/sbin/iptables -D INPUT -m set --match-set countries-blocklist-v4 src -j DROP >/dev/null 2>&1
 /sbin/ipset destroy countries-blocklist-v4 >/dev/null 2>&1
 
 /sbin/ipset create countries-blocklist-v4 hash:net comment
 
 /sbin/ipset restore < "$ripedeny_file"
 
-/sbin/iptables -I NEEDRESTRICT -m set --match-set countries-blocklist-v4 src -j DROP
-#/sbin/iptables -I INPUT -m set --match-set countries-blocklist-v4 src -j DROP
-
+#/sbin/iptables -I NEEDRESTRICT -m set --match-set countries-blocklist-v4 src -j DROP
+/sbin/iptables -I INPUT -m set --match-set countries-blocklist-v4 src -j DROP